### 2024-02-13 #6352 @editorialbot
[PRE REVIEW]: BayesEoR: Bayesian 21-cm Power Spectrum Estimation from Interferometric Visibilities
**Submitting author:** @jburba (<a href="http://orcid.org/0000-0002-8465-9341">Jacob Burba</a>)
**Repository:** https://github.com/PSims/BayesEoR
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
