### 2024-03-22 a95d767
Fixed bug with noise calculation if Ntimes is odd; Added jsonargparse.ArgumentParser to data preprocessing; Deleted unused elementwise average function
