### 2024-03-22 f41fabc
Check if running in SLURM, if so create an empty file with the Job ID as the name in the output directory
