### 2024-03-26 1aad577
Updated file naming syntax to include time info; replace format calls with f strings
