+++ scripts/data_preprocessing.py
@@ -224,9 +224,7 @@ def add_mtime_to_filename(path, filename_in, join_char='-'):
     suffix = fp.suffix
     mtime = datetime.fromtimestamp(os.path.getmtime(fp))
     mtime = mtime.isoformat()
-    filename_out = filename_in.replace(suffix, '{}{}{}'.format(
-        join_char, mtime, suffix
-    ))
+    filename_out = filename_in.replace(suffix, f'{join_char}{mtime}{suffix}')
     return filename_out
 
 
@@ -298,6 +296,7 @@ def data_processing(
         opts,
         filename,
         min_freq_MHz,
+        central_jd,
         save_dir="./",
         inst_model_dir=None,
         uvd_all_bls=None
@@ -322,6 +321,10 @@ def data_processing(
         Dictionary containing command line arguments.
     filename : str
         Input filename of UVData compatible file.
+    min_freq_MHz : float
+        Minimum frequency in the data.
+    central_jd : float
+        Central JD in the data.
     save_dir : str
         Path where the data vector will be written.  Defaults to the `scripts/`
         subdirectory within the installation path of `BayesEoR`.
@@ -333,38 +336,32 @@ def data_processing(
         baselines.
 
     """
-    outfile = filename.replace(
-        '.uvh5',
-        '-start-freq-{:.2f}-nf-{}-nbls-{}'.format(
-            min_freq_MHz, opts.nf, uvd_select.Nbls*2
-        )
-    )
+    # wn-gsm-ptsrc-fov-30-start-freq-150.29-nf-180-nbls-30-pol-pI.npy
+    outfile = filename.replace('.uvh5', '')
+    # Frequency parameters
+    outfile += f'-min-freq-{min_freq_MHz:.2f}MHz-Nfreqs-{opts.nf}'
     if opts.avg_adj_freqs:
-        outfile = outfile.replace(
-            '-nf-{}'.format(opts.nf), '-nf-{}-adj-freq-avg'.format(opts.nf)
-        )
+        outfile += '-adj-freq-avg'
+    # Baseline parameters
+    if opts.bl_type:
+        outfile += f'-Nbls-{uvd_select.Nbls*2}-{opts.bl_type}'
+    else:
+        if uvd_select.Nbls > 1:
+            outfile += f'-Nbls-{uvd_select.Nbls*2}'
+        elif uvd_select.Nbls == 1:
+            outfile += f'-bl-{antnums[0]}-{antnums[1]}'
+    # Time parameters
+    outfile += f'-Ntimes-{opts.nt}-central-JD-{central_jd:.2f}'
+    # Polarization parameters
     if opts.form_pI:
         outfile += '-pol-pI'
     else:
         outfile += f'-pol-{opts.pol}'
+    # Phasing parameters
     if opts.phase:
-            outfile += '-phased'
+        outfile += '-phased'
     elif opts.phase_time_jd is not None:
-        outfile += '-phase-time-{}'.format(opts.phase_time_jd)
-
-    if uvd_select.Nbls == 1:
-        antnums = uvd_select.baseline_to_antnums(uvd_select.baseline_array[0])
-        outfile = outfile.replace(
-            'nbls-{}-'.format(uvd_select.Nbls),
-            'bl-{}-{}-'.format(
-                uvd_select.Nbls, antnums[0], antnums[1]
-            )
-        )
-    elif opts.bl_type:
-        outfile = outfile.replace(
-            'nbls-{}-'.format(uvd_select.Nbls),
-            'nbls-{}-{}-'.format(uvd_select.Nbls, opts.bl_type)
-        )
+        outfile += f'-phase-time-{opts.phase_time_jd}'
     outfile += '.npy'
[…]
