### 2024-03-28 6ec2749
Fix bug in model k-cube generation if Nfreqs is odd
