### 2024-03-31 #6352 @dfm
:wave: @steven-murray, @sterinaldi, @musoke — Would any of you be available and willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
