### 2024-04-03 #6352 @dfm
@steven-murray — Thanks for your response! I'd love to have you review if it makes sense.

The JOSS Conflict of Interest policy is available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy

As you can see, depending on the details of the collaborations and whether or not you feel you can be unbiased, this could be waived. Let me know what you think, and I'm happy to chat more via email if that would be useful. Thanks again!
