### 2024-04-09 923b1bd
GPU interface now in a module bayeseor.gpu; C code now stored in bayeseor/gpu/
