+++ bayeseor/gpu/__init__.py
@@ -1,0 +1,1 @@
+from .gpu import GPU_Interface
