+++ bayeseor/utils/map.py
@@ -1,0 +1,229 @@
+""" Class and functions for performing maximum a posteriori calculations. """
+import numpy as np
+from pathlib import Path
+import json
+from jsonargparse import Namespace
+import matplotlib.pyplot as plt
+
+from ..matrices import BuildMatrixTree
+from ..model import (
+    generate_k_cube_in_physical_coordinates,
+    mask_k_cube,
+    generate_k_cube_model_spherical_binning,
+    calc_mean_binned_k_vals
+)
+from ..params import calculate_derived_params
+from ..posterior import PowerSpectrumPosteriorProbability
+
+class MaximumAPosteriori(object):
+    """
+    Class for performing maximum a posteriori calculations with BayesEoR.
+
+    Parameters
+    ----------
+    data_path : str or Path
+        Path to a numpy readable visibility data file in units of mK sr.
+    array_dir : str or Path
+        Path to a directory containing T, Ninv, and T_Ninv_T.
+    output_dir : str or Path
+        Path to a BayesEoR output directory which contains 'args.json' and
+        'k-vals.txt' files.
+    verbose : bool
+        If True (default), print timing and status messages.
+    
+    """
+    def __init__(
+        self,
+        data_path,
+        array_dir,
+        output_dir,
+        verbose=True
+    ):        
+        if not isinstance(output_dir, Path):
+            output_dir = Path(output_dir)
+        with open(output_dir / "args.json", "r") as f:
+            cl_args = Namespace(**json.load(f))
+        args = calculate_derived_params(cl_args)
+
+        # Calculate k bin values
+        mod_k = generate_k_cube_in_physical_coordinates(
+            args.nu,
+            args.nv,
+            args.neta,
+            args.ps_box_size_ra_Mpc,
+            args.ps_box_size_dec_Mpc,
+            args.ps_box_size_para_Mpc
+        )[0]
+        mod_k_vo = mask_k_cube(mod_k)
+        k_cube_voxels_in_bin = generate_k_cube_model_spherical_binning(
+            mod_k_vo,
+            args.ps_box_size_para_Mpc
+        )[0]
+        k_vals = calc_mean_binned_k_vals(
+            mod_k_vo,
+            k_cube_voxels_in_bin,
+            rank=1-int(verbose)
+        )
+
+        # Load matrices
+        bm = BuildMatrixTree(
+            array_dir.as_posix(),
+            args.include_instrumental_effects,
+            args.use_sparse_matrices
+        )
+        Ninv = bm.read_data(array_dir / "Ninv.npz", "Ninv")
+        T = bm.read_data(array_dir / "T.h5", "T")
+        T_Ninv_T = bm.read_data(array_dir / "T_Ninv_T.h5", "T_Ninv_T")
+
+        d = np.load(data_path)
+        Ninv_d = np.dot(Ninv, d)
+        d_Ninv_d = np.dot(d.conjugate(), Ninv_d)
+        dbar = np.dot(T.conjugate().T, Ninv_d)
+        
+        self.pspp = PowerSpectrumPosteriorProbability(
+            T_Ninv_T,
+            dbar,
+            k_vals,
+            k_cube_voxels_in_bin,
+            args.nuv,
+            args.neta,
+            args.nf,
+            args.nq,
+            Ninv,
+            d_Ninv_d,
+            args.redshift,
+            args.ps_box_size_ra_Mpc,
+            args.ps_box_size_dec_Mpc,
+            args.ps_box_size_para_Mpc,
+            include_instrumental_effects=args.include_instrumental_effects,
+            log_priors=args.log_priors,
+            uprior_inds=args.uprior_inds,
+            inverse_LW_power=args.inverse_LW_power,
[…]
