### 2024-04-10 1499dd1
Missed class rename in __init__
