### 2024-04-12 c177e80
Add noise generation, Hermitian checks, and add more attributes to MaximumAPosteriori class on init
