+++ bayeseor/utils/map.py
@@ -7,13 +7,16 @@ import matplotlib.pyplot as plt
 
 from ..matrices import BuildMatrixTree
 from ..model import (
+    load_inst_model,
     generate_k_cube_in_physical_coordinates,
     mask_k_cube,
     generate_k_cube_model_spherical_binning,
-    calc_mean_binned_k_vals
+    calc_mean_binned_k_vals,
+    generate_data_and_noise_vector_instrumental
 )
-from ..params import calculate_derived_params
+from ..params import calculate_derived_params, parse_uprior_inds
 from ..posterior import PowerSpectrumPosteriorProbability
+from .utils import mpiprint, vector_is_hermitian
 
 class MaximumAPosteriori(object):
     """
@@ -46,6 +49,9 @@ class MaximumAPosteriori(object):
         args = calculate_derived_params(cl_args)
 
         # Calculate k bin values
+        mpiprint(
+            "\nSpherical k bins:", style="bold", rank=1-int(verbose)
+        )
         mod_k = generate_k_cube_in_physical_coordinates(
             args.nu,
             args.nv,
@@ -59,38 +65,149 @@ class MaximumAPosteriori(object):
             mod_k_vo,
             args.ps_box_size_para_Mpc
         )[0]
-        k_vals = calc_mean_binned_k_vals(
+        self.k_vals = calc_mean_binned_k_vals(
             mod_k_vo,
             k_cube_voxels_in_bin,
             rank=1-int(verbose)
         )
 
         # Load matrices
-        bm = BuildMatrixTree(
+        self.bm = BuildMatrixTree(
             array_dir.as_posix(),
             args.include_instrumental_effects,
             args.use_sparse_matrices
         )
-        Ninv = bm.read_data(array_dir / "Ninv.npz", "Ninv")
-        T = bm.read_data(array_dir / "T.h5", "T")
-        T_Ninv_T = bm.read_data(array_dir / "T_Ninv_T.h5", "T_Ninv_T")
+        Ninv = self.bm.read_data((array_dir / "Ninv.npz").as_posix(), "Ninv")
+        self.Ndata = Ninv.shape[0]
+        T = self.bm.read_data((array_dir / "T.h5").as_posix(), "T")
+        T_Ninv_T = self.bm.read_data(
+            (array_dir / "T_Ninv_T.h5").as_posix(), "T_Ninv_T"
+        )
 
-        d = np.load(data_path)
-        Ninv_d = np.dot(Ninv, d)
-        d_Ninv_d = np.dot(d.conjugate(), Ninv_d)
+        data = np.load(data_path, allow_pickle=True)
+        if data.dtype.kind == "O":
+            # Data and noise vector saved in a single dictionary
+            dict_format = True
+            data = data.item()
+            if "noise" in data and "sigma" not in args:
+                gen_noise = False
+            else:
+                gen_noise = True
+        else:
+            # Data and noise vector saved as separate arrays
+            dict_format = False
+            if "noise_data_path" in args:
+                gen_noise = False
+            else:
+                gen_noise = True
+        if gen_noise:
+            effective_noise = None
+        else:
+            if dict_format:
+                effective_noise = data["noise"]
+            else:
+                effective_noise = np.load(args.noise_data_path)
+            args.sigma = effective_noise.std()
+        if args.include_instrumental_effects:
+            uvw_array_m, bl_red_array, phasor_vec = load_inst_model(
+                args.inst_model
+            )
+            if phasor_vec is not None and args.drift_scan:
+                phasor_vec = None
+            bl_red_array = bl_red_array*0 + bl_red_array.min()
+            avg_bl_red = np.mean(bl_red_array)
+            if gen_noise:
+                args.sigma *= avg_bl_red**0.5
+            if dict_format:
[…]
