### 2024-04-15 #6352 @sibirrer
> 👋 @paddyroddy, @sibirrer, @jrs65 — Would any of you be available and willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @dfm, I am currently overwhelmed with other grants and reviewing duties and hence I decline this time. Apologies
