### 2024-04-16 8b5267f
Instrument model written as individual files, not dictionary; Updated add_mtime_to_filename -> add_mtime_to_path to support files and directories; Updated load_inst_model variable names and docstring
