### 2024-04-23 7d84163
First push of new file_root function
Deleted `generate_output_file_base` and replaced it and the `file_root` generating lines in `run-analysis.py` with a new function `gen_output_file_base`.
