### 2024-04-24 PSims/BayesEoR#20 @zonca
[joss-review] how to create `wrapmzpotrf.so`
There are no instructions on how to build `wrapmzpotrf.so`:

https://bayeseor.readthedocs.io/en/latest/installation.html#shared-library-creation

the best would be to add this to `pyproject.toml` as a step in the installation of the Python package. However, if that is too complicated, it would be useful to provide an example Makefile on how to compile the shared library.
