### 2024-04-24 #6667 @zonca
@jburba I opened issues on the repository with some feedback, the only one which is blocking is https://github.com/PSims/BayesEoR/issues/20, if you focus first on this issue, I can continue trying to run the example config. I was able to build the matrix stack, but cannot run sampling.
