### 2024-04-24 a031b54
fix: this flag is required by  in Ubuntu 22.04
