+++ bayeseor/gpu/wrapmzpotrf.c
@@ -1,4 +1,4 @@
-
+#define ADD_ // MAGMA Fortran name mangling parameter
 #include <stdio.h>
 #include <stdlib.h>
 #include "cublas_v2.h"     // if you need CUBLAS, include before magma.h
