### 2024-04-24 87f31b7
feat: simple Makefile to build wrapmzpotrf.so
tested on Ubuntu 22.04 with MAGMA from repositories
