### 2024-04-24 02bee94
feat: use CONDA_PREFIX in Makefile
