### 2024-04-24 90d1c32
feat: install magma and cuda with conda
tested on Ubuntu 22.04 with A100 (Jetstream 2)
