+++ environment.yaml
@@ -20,3 +20,5 @@ dependencies:
   - setuptools
   - setuptools_scm
   - sphinx
+  - cuda
+  - magma
