### 2024-04-24 PSims/BayesEoR#20 @zonca
see https://github.com/PSims/BayesEoR/pull/23
