### 2024-04-25 #6667 @musoke
@jburba, would you like to respond to some of  @zonca's requests before I finish my review?  As is, I expect that many of my comments will be agreement that installation should be simplified/clarified.

(@dfm, let me know if this would be procedurally problematic 😃)
