### 2024-04-26 #6667 @dfm
Thanks all!

@musoke — I'd probably recommend that you see how far you can get with your review, as long as @zonca's issues aren't full blockers (it does seem like the installation one might be?). That way @jburba could work on addressing all the comments in parallel. Otherwise this could end up dragging on longer than we want. But you are all welcome to proceed as you feel is best for everyone involved!
