### 2024-04-30 cba04da
Install magma libraries from conda-forge
Look for lib/libmagma.so instead of a user-compiled .so file in the bayeseor/gpu/ directory.  First attempt at updating the GPUInterface class to use the conda-forge magma libs.
