+++ bayeseor/gpu/gpu.py
@@ -1,5 +1,6 @@
 import numpy as np
 from pathlib import Path
+from sysconfig import get_paths
 
 from ..utils import mpiprint
 
@@ -20,7 +21,8 @@ class GPUInterface(object):
     """
     def __init__(self, base_dir=None, rank=0, verbose=True):
         if base_dir is None:
-            self.base_dir = Path(__file__).parent
+            # Look for MAGMA .so files in environment's lib directory
+            self.base_dir = Path(get_paths()['stdlib']).parent
         else:
             self.base_dir = Path(base_dir)
         self.rank = rank
@@ -32,29 +34,35 @@ class GPUInterface(object):
             import ctypes
             from numpy import ctypeslib
 
-            so_path = self.base_dir / 'wrapmzpotrf.so'
+            so_path = self.base_dir / 'libmagma.so'
             if self.verbose:
                 mpiprint(
                     f'Loading shared library from {so_path}',
                     rank=self.rank
                 )
-            # wrapmzpotrf contains a python wrapper function of the C routine
-            # from the Matrix Algebra for GPU and Multicore Architectures
-            # (MAGMA) library called mzpotrf which computes the Cholesky
-            # decomposition of a complex, Hermitian, positive-definite matrix.
-            wrapmzpotrf = ctypes.CDLL(so_path)
-            # nrhs specifies the number of columns in the object on the right
-            # hand side of the linear system Ax=b.  In our case, b is a column
-            # vector and thus nrhs = 1 (1 column).
-            self.nrhs = 1
-            wrapmzpotrf.cpu_interface.argtypes = [
-                ctypes.c_int,
+            # libmagma.so contains functions from the Matrix Algebra for GPU
+            # and Multicore Architectures (MAGMA) library.  We are interested
+            # in the function magma_zpotrf from include/magma_z.h which
+            # computes the Cholesky decomposition of a complex, Hermitian,
+            # positive-definite matrix.  Please see the MAGMA documentation
+            # linked in the BayesEoR documentation for more information.
+            magma = ctypes.CDLL(so_path)
+            self.magma_zpotrf = magma.magma_zpotrf
+            # The function magma_zpotrf takes as arguments
+            # 1. uplo (str): 'L' or 'U' for lower or upper triangular
+            #                decomposition of dA is stored
+            # 2. n (int): The order of the matrix dA
+            # 3. dA (complex pointer): pointer to the matrix dA with shape
+            #                          (ldda, n)
+            # 4. ldda (int): The leading dimension of dA
+            # 5. info (int pointer): Info flag
+            self.magma_zpotrf.argtypes = [
+                ctypes.c_wchar_p,
                 ctypes.c_int,
                 ctypeslib.ndpointer(np.complex128, ndim=2, flags='C'),
-                ctypeslib.ndpointer(np.complex128, ndim=1, flags='C'),
                 ctypes.c_int,
-                ctypeslib.ndpointer(int, ndim=1, flags='C')]
-            self.wrapmzpotrf = wrapmzpotrf
+                ctypeslib.ndpointer(int, ndim=1, flags='C')
+            ]
             if self.verbose:
                 mpiprint('Computing on GPU(s)', rank=self.rank, end='\n\n')
             self.gpu_initialized = True
