### 2024-04-30 20ce658
Use zpotrf from libmagma.so in gpu and posterior modules; Add init and finalize functions to gpu and posterior modules
Uses an int for the first parameter to magma_zpotrf which matches the magma type definition for `magma_uplo_t` in `include/magma_types.h`.  I incorrectly used a string in the last commit.  Add magma_init and magma_finalize functions which we need to call before and after magma_zpotrf, respectively.  Updated the posterior module to use these new functions in place of the old wrapper functino.
