### 2024-05-01 9a3f5e1
Fix magma package name
