### 2024-05-01 9b7baab
Update installation instructions to use conda's magma
