+++ docs/installation.rst
@@ -11,15 +11,17 @@ BayesEoR relies on GPUs to perform a Cholesky decomposition on large matrices us
 - MPI
 - `MultiNest <https://github.com/JohannesBuchner/MultiNest>`_
 
+All of these dependcies can be installed via ``conda`` for ease of use.  Please see the section below on :ref:`python-dependencies` for more information.
+
 BayesEoR has been succesfully run with
 
 - GPUs: NVIDIA P100, V100, and A100 architectures
-- MAGMA: 2.4.0, 2.5.4, and 2.7.1
+- MAGMA: ``conda`` and source installations
 - MPI: ``conda`` installation (mpich) and OpenMPI 4.0.5
 - CUDA: 9.1.85.1 and 11.1.1
-- MultiNest: ``conda`` installation and a source installation
+- MultiNest: ``conda`` and source installations
 
-This is not an exhaustive list of software versions which are compatible with our analysis, just a guide of what versions we have used succesfully in our BayesEoR analyses.
+This is not an exhaustive list of software versions which are compatible with our analysis, just a guide of what versions we have used succesfully in our own analyses.
 
 
 
@@ -29,39 +31,20 @@ A Note on Using CPUs
 While it is in principle possible to run BayesEoR on CPUs, we strongly suggest using GPUs due to their increased speed and precision relative to CPU-based methods.
 
 
-
-MAGMA Install Notes
-"""""""""""""""""""
-
-We have found that MAGMA performs best when used with Intel's math kernel library (`MKL <https://www.intel.com/content/www/us/en/developer/tools/oneapi/onemkl-download.html>`_) when running on Intel-based machines.  It is also possible to build MAGMA with `OpenBLAS <https://www.openblas.net/>`_.  For more configuration options when installing MAGMA, consult the README included with your source download.  Below, we also provide some installation notes that proved useful when installing MAGMA.
-
-- `Download <https://icl.cs.utk.edu/magma/>`_ the latest version (2.7.1 at the time of this writing) and checkout README for some simple quick start instructions.  Upon extracting the files, this makes a directory `magma-<version>/`.
-- There are a variety of example make.inc files in the `magma-<version>/make.inc-examples` directory.  For reference, we have used the `make.inc.mkl-gcc` file.  Copy whichever example make file you'd like to use to `magma-<version>/make.inc`.
-- If you're using ``CUDA`` and ``MKL``, make sure to uncomment the lines in the `make.inc` file that set the ``MKLROOT`` and ``CUDADIR`` bash environment variables.  Otherwise, uncomment the appropriate lines for the libraries you are using.  These lines are commented by default.  Set these variables to their appropriate paths on your machine.
-- We also found it necessary to append the path to our installed ``CUDA`` binaries to the bash ``PATH`` variable, in our case by executing ``export PATH=/usr/local/cuda/bin:$PATH``, because the installer requires use of NVIDIA's ``nvcc`` compiler.
-- Inside the `Makefile`, set the prefix variable to install ``MAGMA`` to your desired location.
-- Install
-- Note that running ``make`` might take a while.  We have seen ``make`` take as long as 1-1.5 hours.  Consider running ``make`` in a screen instance to avoid network issues.
-
-
-
-Shared Library Creation
-^^^^^^^^^^^^^^^^^^^^^^^
-
-Included within our repo is a file containing the MAGMA calls written in C.  This C code must be compiled into a shared library and placed inside the directory `bayeseor/posterior/gpu_wrapper/`.  BayesEoR automatically looks for a shared library in this directory with name `wrapmzpotrf.so`.
-
-
+.. _python-dependencies:
 
 Python Dependencies
 -------------------
 
[…]
