### 2024-05-03 6577972
Add cuda as an optional dependency
Add cuda to environment.yaml and docs/installation so cuda is nominally installed via conda
