### 2024-05-07 0acf77f
Add RAM requirements for test data runs to docs/usage
