+++ docs/usage.rst
@@ -72,13 +72,13 @@ Test Dataset
 
 The BayesEoR repository provides a set of test data and an example yaml configuration file.  The test data contain mock EoR only simulated visibilities with a Gaussian beam and a full width at half maximum of 9.3 degrees.  For more information on the test data, see Section 3 of `Burba et al. 2023 <https://ui.adsabs.harvard.edu/abs/2023MNRAS.520.4443B/abstract>`_.
 
-To build the matrices (which will require ~17 GB of disk space) using the provided example configuration yaml and test data, first navigate to the root directory of the BayesEoR repository and run
+To build the matrices (which will require ~20 GB of RAM ~17 GB of disk space) using the provided example configuration yaml and test data, first navigate to the root directory of the BayesEoR repository and run
 
 .. code-block:: Bash
 
     python run-analysis.py --config example-config.yaml --cpu
 
-Once the matrices are built, you can run the power spectrum analysis via
+Once the matrices are built, you can run the power spectrum analysis (which will require ~12 GB of RAM) via
 
 .. code-block:: Bash
 
