### 2024-05-07 b9af6b5
More accurate estimate for matrix construction RAM
