### 2024-05-07 f187521
First push of paper updates to address #25
- Added a sentence where the 21-cm emission comes from
- Replaced "Gyr" with billion years
- Added a paragraph to the summery specifying the components of the instrument model, the form of the input data (with an added citation to pyuvdata), and some examples of analysis parameters
- Updated the figure caption to define the symbols in the subplot labels (\Delta^2(k), k, and \varphi)
