+++ joss/paper.md
@@ -35,15 +35,17 @@ bibliography: paper.bib
 
 # Statement of need
 
-The highly redshifted 21-cm signal from neutral hydrogen in the early universe contains a wealth of information about the state of the intergalactic medium during the first Gyr of cosmic history from which properties of the first stars and galaxies can be inferred. Modern interferometers like HERA (http://reionization.org/), LoFAR (https://www.mpifr-bonn.mpg.de/en/lofar), and the MWA (http://www.mwatelescope.org/), have been designed to observe with many antennas simultaneously to maximize their sensitivity to the 21-cm signal. These experiments have shown that detecting this signal is rife with difficulty [@hera:2022; @lofar:2020; @mwa:2020], primarily due to the presence of bright, contaminating sources between us and the cosmological signal, referred to as "foregrounds." Because these foregrounds are several orders of magnitude brighter than the 21-cm signal we wish to detect, advanced data analysis tools are needed to extract accurate estimates of the faint, 21-cm signal.
+The 21-cm signal from neutral hydrogen is emitted as a result of a spin-flip transition in which the quantum spins of the proton and electron transition from an aligned to an anti-aligned state. By observing this signal at high redshift in the early universe, we can learn a wealth of information about the state of the intergalactic medium during the first billion years of cosmic history from which properties of the first stars and galaxies can be inferred. Modern interferometers like HERA (http://reionization.org/), LoFAR (https://www.mpifr-bonn.mpg.de/en/lofar), and the MWA (http://www.mwatelescope.org/), have been designed to observe with many antennas simultaneously to maximize their sensitivity to the 21-cm signal. These experiments have shown that detecting this signal is rife with difficulty [@hera:2022; @lofar:2020; @mwa:2020], primarily due to the presence of bright, contaminating sources between us and the cosmological signal, referred to as "foregrounds." Because these foregrounds are several orders of magnitude brighter than the 21-cm signal we wish to detect, advanced data analysis tools are needed to extract accurate estimates of the faint, 21-cm signal.
 
 # Summary
 
 `BayesEoR` is a GPU-accelerated, Python-based implementation of a Bayesian framework designed to jointly model the 21-cm and foreground signals and forward model the instrument with which these signals are observed. Using these combined techniques, we can overcome the aforementioned difficulties associated with extracting a faint, background signal in the presence of bright foregrounds. `BayesEoR` enables one to sample directly from the marginal posterior distribution of the power spectrum of the underlying 21-cm signal in interferometric data, enabling recovery of statistically robust and unbiased[^1] estimates of the 21-cm power spectrum and its uncertainties [@sims:2016; @sims:2019a; @sims:2019b; @burba:2023]. The outputs of an analysis using the [test dataset](https://bayeseor.readthedocs.io/en/latest/usage.html#test-dataset) and [plotting code](https://bayeseor.readthedocs.io/en/latest/usage.html#analyzing-bayeseor-outputs) provided with `BayesEoR` can be found in \autoref{fig:example}.
 
[…]
