### 2024-05-07 2e74011
Add more information about running BayesEoR for #25
