### 2024-05-09 992ac3a
Add Peter's suggestions
1. Add a sentence mentioning the EoR in the statement of need to make the choice of name (
BayesEoR) more obvious.
2. Fix LoFAR -> LOFAR
3. Added a missing space to fix "theMWA"
4. Reword the last two sentences in the statement of need to highlight the covariance of the _observed_ EoR and foreground signals which better motivates the forward model approach.
5. Added a couple sentences regarding memory requirements for smaller test cases to the summary.
6. Added a comma in the figure caption when defining $k$
