+++ joss/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-Neutral hydrogen can undergo a spin-flip transition in which the quantum spins of the proton and electron transition from an aligned to an anti-aligned state, or vice versa, resulting in emission or absorption of a photon with a wavelength of 21-cm. The hydrogen 21-cm spin temperature quantifies the relative number densities of atoms in the aligned and anti-aligned states. Interferometric 21-cm cosmology experiments aim to measure the contrast between the 21-cm spin temperature of neutral hydrogen and the radio background temperature in the early Universe. By observing this signal at high redshift, we can learn a wealth of information about the state of the intergalactic medium during the first billion years of cosmic history.  This information can, in turn, be used to infer properties of the first stars and galaxies that transformed the hydrogen intergalactic medium from a cold neutral gas to a hot ionised plasma during the Epoch of Reionization (EoR). Modern interferometers like [HERA](http://reionization.org/), [LOFAR](https://www.mpifr-bonn.mpg.de/en/lofar), and the [MWA](http://www.mwatelescope.org/), have been designed to observe with many antennas simultaneously to maximize their sensitivity to the 21-cm signal from the EoR. These experiments have shown that detecting this signal is rife with difficulty [@hera:2022; @lofar:2020; @mwa:2020].  This is primarily due to the coupling of bright contaminating sources between us and the cosmological signal, referred to as "foregrounds", with the spectral structure imparted by the instrument. Existing approaches to recovering the 21-cm signal from the data lack direct modelling of the observed covariance between the 21-cm and foreground signals in the data.  Intrinsically, the 21-cm and foregrounds signals are uncorrelated.  The instrument modulates both signals identically during observation, however, making them covariant. This covariance can be accounted for by forward modelling both signals, a key advantage of our approach in BayesEoR.
+Neutral hydrogen can undergo a spin-flip transition in which the quantum spins of the proton and electron transition from an aligned to an anti-aligned state, or vice versa, resulting in emission or absorption of a photon with a wavelength of 21-cm. The hydrogen 21-cm spin temperature quantifies the relative number densities of atoms in the aligned and anti-aligned states. Interferometric 21-cm cosmology experiments aim to measure the contrast between the 21-cm spin temperature of neutral hydrogen and the radio background temperature in the early Universe. By observing this signal at high redshift, we can learn a wealth of information about the state of the intergalactic medium during the first billion years of cosmic history.  This information can, in turn, be used to infer properties of the first stars and galaxies that transformed the hydrogen intergalactic medium from a cold neutral gas to a hot ionised plasma during the Epoch of Reionization (EoR). Modern interferometers like [HERA](http://reionization.org/), [LOFAR](https://www.mpifr-bonn.mpg.de/en/lofar), and the [MWA](http://www.mwatelescope.org/), have been designed to observe with many antennas simultaneously to maximize their sensitivity to the 21-cm signal from the EoR. These experiments have shown that detecting this signal is rife with difficulty [@hera:2022; @lofar:2020; @mwa:2020].  This is primarily due to the coupling of bright contaminating sources between us and the cosmological signal, referred to as "foregrounds", with the spectral structure imparted by the instrument. Existing approaches to recovering the 21-cm signal from the data lack direct modelling of the observed covariance between the 21-cm and foreground signals in the data.  Intrinsically, the 21-cm and foreground signals are uncorrelated.  The instrument modulates both signals identically during observation, however, making them covariant. This covariance can be accounted for by forward modelling both signals, a key advantage of our approach in BayesEoR.
 
 # Summary
 
