### 2024-05-10 #6667 @zonca
Good job @jburba, my review is finished,
there are 3 open issues that I think would be good to address, but I do not think they are blocking.

@musoke I think you can get started, you might want to check the 3 issues I left open in the repo to see if you feel those should be addressed before publication or not: https://github.com/PSims/BayesEoR/issues/created_by/zonca
