### 2024-05-31 PSims/BayesEoR#30 @musoke
[joss-review] Feedback on draft
https://github.com/openjournals/joss-reviews/issues/6667

The paper is much improved since the initial submission :smile:

These comments are in reference to [the latest draft](https://github.com/PSims/BayesEoR/actions/runs/9034432698) ([paper.pdf](https://github.com/user-attachments/files/15519170/paper.pdf)).

#### Writing
- [x] The balance between summary and statement of need seems off.
The [JOSS docs](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) ask for "A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience." I was expecting the summary to be similar to an abstract:  first and shorter.

- [x] I would suggest adding sentences similar to the current lines 16-20 (about HERA/LOFAR/MWA and what they hope to learn) and 31-32 (your approach) to the start.  Only explain details of 21cm after convincing the reader that it is interesting.
- [x] Figure: Say explicitly that inferred parameters on the left are the same coefficients as on the right.
- [x] Explain the quantities in the figure.  Is this a power spectrum of 21-cm signal? Or hydrogen density?
- [x] Give an equation for $\Delta^2(k) = \sum ...$ as well as the current textual descritption.

#### State of the field
- [x] You briefly explain why this is a difficult task and describe your statistical techniques and implementation.  I would like to see some mention of other approaches and how BayesEoR differs.
- [x] Is there alternative software that can be cited?  Perhaps from the HERA, LoFAR, or MWA collaborations.

#### References
- [x] Should MultiNest be cited?  It seems central to the analysis technique.
