### 2024-06-04 PSims/BayesEoR#31 @musoke
[joss-review] Error when running tests
https://github.com/openjournals/joss-reviews/issues/6667

I have installed the package and attempted to run the tests described in the docs (https://bayeseor.readthedocs.io/en/latest/usage.html#test-dataset)

Running
```
python run-analysis.py --config example-config.yaml --cpu
python run-analysis.py --config example-config.yaml --gpu
```
with the default  `example-config.yaml` results in an error:
```
(bayeseor) [nathan@host BayesEoR] $ python run-analysis.py --config example-config.yaml --gpu

mpi_size: 1

╭────────────────────────────────────────────────────────────────────────────────────────────────────────╮
│ Parameters                                                                                             │
╰────────────────────────────────────────────────────────────────────────────────────────────────────────╯
{'achromatic_beam': False,
 'antenna_diameter': None,
 'array_dir_prefix': './array-storage/',
 'bandwidth_MHz': 9.029521500604,
 'beam_center': None,
 'beam_peak_amplitude': 1.0,
 'beam_ref_freq': None,
 'beam_type': 'gaussian',
 'beta': [2.63, 2.82],
 'central_jd': 2458098.3065661727,
 'channel_width_MHz': 0.237618986858,
 'clobber': False,
 'config': [Path_fr(example-config.yaml, cwd=/home/nathan/src/BayesEoR)],
 'cosfreq': None,
 'data_path': Path_fr(./test_data/visibilities.npy, cwd=/home/nathan/src/BayesEoR),
 'deta': 1.1074783973138646e-07,
 'drift_scan': True,
 'du_eor': 4.438755506838748,
 'du_fg': 4.438755506838748,
 'dv_eor': 4.438755506838748,
 'dv_fg': 4.438755506838748,
 'eor_random_seed': 892736,
 'eor_sim_path': None,
 'file_root': None,
 'fit_for_monopole': False,
 'fit_for_shg_amps': False,
 'fit_for_spectral_model_parameters': False,
 'fov_dec_eor': 12.9080728652,
 'fov_dec_fg': 12.9080728652,
 'fov_ra_eor': 12.9080728652,
 'fov_ra_fg': 12.9080728652,
 'freqs_MHz': array([158.30404874, 158.54166773, 158.77928672, 159.0169057 ,
       159.25452469, 159.49214368, 159.72976266, 159.96738165,
       160.20500064, 160.44261962, 160.68023861, 160.9178576 ,
       161.15547659, 161.39309557, 161.63071456, 161.86833355,
       162.10595253, 162.34357152, 162.58119051, 162.81880949,
       163.05642848, 163.29404747, 163.53166645, 163.76928544,
       164.00690443, 164.24452341, 164.4821424 , 164.71976139,
       164.95738038, 165.19499936, 165.43261835, 165.67023734,
       165.90785632, 166.14547531, 166.3830943 , 166.62071328,
       166.85833227, 167.09595126]),
 'fwhm_deg': 9.306821090681533,
 'include_instrumental_effects': True,
 'inst_model': Path_dr(./test_data/, cwd=/home/nathan/src/BayesEoR),
 'integration_time_seconds': 11.0,
 'inverse_LW_power': 1e-16,
 'log_priors': True,
 'neta': 38,
 'nf': 38,
 'noise_data_path': None,
 'noise_seed': 742123,
 'npl': 2,
 'npl_sh': None,
 'nq': 0,
 'nq_sh': None,
 'nside': 128,
 'nt': 34,
 'nu': 15,
 'nu_fg': 15,
 'nu_min_MHz': 158.304048743,
 'nu_sh': None,
 'nuv': 224,
 'nuv_fg': 224,
 'nuv_sh': None,
 'nv': 15,
 'nv_fg': 15,
 'nv_sh': None,
 'output_dir': './chains/',
 'pl_grid_spacing': None,
 'pl_max': None,
 'pl_min': None,
 'priors': [[-2.0, 2.0],
            [-1.2, 2.8],
            [-0.7, 3.3],
            [0.7, 2.7],
            [1.1, 3.1],
            [1.5, 3.5],
            [2.0, 4.0],
            [2.4, 4.4],
            [2.7, 4.7]],
 'ps_box_size_dec_Mpc': 2039.0643396379605,
 'ps_box_size_para_Mpc': 148.80444173068133,
 'ps_box_size_ra_Mpc': 2039.0643396379605,
 'redshift': 7.7302135941678465,
 'sigma': 0.00615864342588761,
 'simple_za_filter': True,
 'single_node': False,
 'speed_of_light': 299792458.0,
 'taper_func': None,
 'telescope_latlonalt': [-30.72152777777791,
                         21.428305555555557,
                         1073.0000000093132],
 'uprior_bins': '',
 'useGPU': True,
 'use_LWM_Gaussian_prior': False,
 'use_Multinest': True,
 'use_intrinsic_noise_fitting': False,
 'use_shg': False,
 'use_sparse_matrices': True,
 'verbose': False}

Output directory: /home/nathan/src/BayesEoR/chains/MN-Test-15-15-38-0-2-6.2E-03-2.63-2.82-lp-dPS-v1

╭────────────────────────────────────────────────────────────────────────────────────────────────────────╮
│ Matrices                                                                                               │
╰────────────────────────────────────────────────────────────────────────────────────────────────────────╯
Array save directory: array-storage/nu-15-nv-15-neta-38-sigma-6.16E-03-nside-128-fov-deg-12.9-za-filter-nq-0/test_data-gaussian-beam-fwhm-9.3068deg-dspb/
Matrix stack complete

---Calculating k-vals---
0 0.04618764702082546
1 0.08652193815249166
2 0.12806815323513662
3 0.16994735208289397
4 0.23300561988455756
5 0.3172473331596665
6 0.44376262151460144
7 0.6125469140767436
8 0.7518323962451139

╭────────────────────────────────────────────────────────────────────────────────────────────────────────╮
│ Data and Noise                                                                                         │
╰────────────────────────────────────────────────────────────────────────────────────────────────────────╯

Using data at ./test_data/visibilities.npy

Generating noise:
Seeding numpy.random with 742123

Hermitian symmetry checks:
signal is Hermitian: True
signal + noise is Hermitian: True

SNR:
Stddev(signal) = 3.0068e-03
Stddev(noise) = 6.1603e-03
SNR = 4.8809e-01

╭────────────────────────────────────────────────────────────────────────────────────────────────────────╮
│ Posterior                                                                                              │
╰────────────────────────────────────────────────────────────────────────────────────────────────────────╯
priors = [[-2.0, 2.0], [-1.2, 2.8], [-0.7, 3.3], [0.7, 2.7], [1.1, 3.1], [1.5, 3.5], [2.0, 4.0], [2.4, 4.4], [2.7, 4.7]]

Instantiating posterior class:
Using log-priors
Calculating dimensionless_PS
Setting inverse_LW_power to 1e-16
Loading shared library from /home/nathan/anaconda3/envs/bayeseor/lib/libmagma.so
Computing on GPU(s)

                    MPI size == 1, analysis will only be run with --single-node flag.

Skipping sampling, exiting...
```

The contents of `./chains/ is
```
(bayeseor) [nathan@larb BayesEoR] $ tree chains/
chains/
└── MN-Test-15-15-38-0-2-6.2E-03-2.63-2.82-lp-dPS-v1
    ├── k-vals-bins.txt
    ├── k-vals-nsamples.txt
    └── k-vals.txt

1 directory, 3 files
```
