### 2024-06-05 #6667 @musoke
I have finished [my review](https://github.com/openjournals/joss-reviews/issues/6667#issuecomment-2072423063).

Summary: it's a nice package but there are some outstanding issues to address.
- https://github.com/PSims/BayesEoR/issues/30
- https://github.com/PSims/BayesEoR/issues/31 I don't doubt that the package does what it claims, but this issue prevented me from verifying the functionality.
