### 2024-06-10 695e361
Run example analysis with a single MPI process
Add `single_node: True` to the configuration yaml so the power spectrum analysis will be run even if run with a single MPI process, i.e. even if `mpi_size = 1`.
