### 2024-06-10 PSims/BayesEoR#31 @jburba
Hi @musoke , thanks for pointing this out.  I've added another flag to the `example-config.py` file which will run the analysis if you're using a single MPI process in my latest commit [695e361](https://github.com/PSims/BayesEoR/commit/695e36191184c0d8351675089f16b88c13dfe1e9).  That toggle is controlled by the `--single-node` command line argument or `single_node` in the configuration yaml.  Please try running
```
python run-analysis.py --config example-config.yaml --gpu
```
again after pulling the latest version of main.
