### 2024-06-10 f8f22e8
Restructure paper and add MultiNest citation
- Restructures the balance between the summary and statement of need sections.
- Adds a new section title for details on running the code as it didn't fall neatly into either "summary" or "statement of need."
- MultiNest has been added as a citation in the summary in addition to a sentence about another useful aspect of BayesEoR, model selection.
- Adds an equation and description of the dimentionless power spectrum
- Adds a phrase to the figure caption saying explicitly the posteriors in the right subplots represent the dimensionless power spectrum
[…]
