### 2024-06-10 72ae556
Update power spectrum equations and figure caption
- Split equations for P(k) and \Delta^2(k)
- Add 1\sigma uncertanties to figure caption
- Rename "Running an analysis" to "Running `BayesEoR`"
