### 2024-06-11 17d185e
Add mention of MPI and cite sims/burba for comparison with other methods
