### 2024-06-11 PSims/BayesEoR#30 @jburba
Hi @musoke , thanks for your suggestions.  I think I've implemented them all in the latest draft.  Please see the latest PDF and let me know what you think.
