### 2024-06-23 #6667 @dfm
@musoke, @zonca — It looks like @jburba has responded to the issues that you opened on the BayesEoR repo. Can you take a look soon and update your checklists / respond to those comments accordingly soon? Thanks!!
