### 2024-06-27 PSims/BayesEoR#31 @jburba
@musoke I've never seen this error before.  A few questions for you:
1. Are you using the latest version of the main branch?
2. Have you created a conda environment from the `environment.yaml` file on the latest main branch?
   -  It looks like you've installed `magma` with conda already, I just want to double check.
3. What system and GPU(s) are you running this test with?

The [function](https://github.com/PSims/BayesEoR/blob/a80c1f2c368353e8354ddd14a3a43764d5eddfee/bayeseor/posterior/posterior.py#L600) in the posterior calculation which uses `magma` already contains a `magma_init()` call, so I think this error suggests a communication error between `magma` and your GPU(s).
