### 2024-06-27 PSims/BayesEoR#31 @musoke
1. Yes.
```
(bayeseor) [nathan@larb BayesEoR] $ git show
commit a80c1f2c368353e8354ddd14a3a43764d5eddfee (HEAD -> main, origin/main, origin/HEAD)
Merge: 695e361 17d185e
Author: jburba <jburba@users.noreply.github.com>
Date:   Tue Jun 11 11:04:33 2024 +0100

    Merge pull request #33 from PSims/joss-edits

    Joss edits to address #30

```

2. I did not recreate the conda environment after pulling the updates, but will try that now.
3. Ubuntu 20.04 (yes, I need to update) with a NVIDIA RTX A3000
```
(bayeseor) [nathan@larb BayesEoR] $ uname -a
Linux larb 5.15.0-107-generic #117~20.04.1-Ubuntu SMP Tue Apr 30 10:35:57 UTC 2024 x86_64 x86_64 x86_64 GNU/Linux
```
and
```
(bayeseor) [nathan@larb BayesEoR] $ nvidia-smi -L
GPU 0: NVIDIA RTX A3000 Laptop GPU (UUID: GPU-e990a63d-4550-5bf1-f637-e85571910ab0)
```

Yes, the error seems to be connected to how/where I ran it, given that it doesn't happen on other machines.
