### 2024-06-27 PSims/BayesEoR#31 @jburba
Hopefully rebuilding your conda environment with the latest environment yaml will help.  We have cuda and magma installed  by conda now.  Depending upon when you built your environment last that may or may not have been the case.

Regardless, please let me know how things go after you rebuild the environment!
