### 2024-06-27 PSims/BayesEoR#31 @jburba
I'm not sure.  I just wanted to make sure both cuda and magma were installed via conda.  I'm trying to dig into the `magma` source now to see if I can find where that error message is triggered.
