### 2024-06-27 PSims/BayesEoR#31 @jburba
I've found a github issue for [pytorch about this error message](https://github.com/pytorch/pytorch/issues/60175) but there isn't a clear answer as to what fixed the error on their end.  It could possibly be related to a version mismatch between `magma` and `cuda`, but this pytorch issue is a few years old so it's hard to say if that's still a possibility.

This [google groups post](https://groups.google.com/a/icl.utk.edu/g/magma-user/c/7j-GI3uzNpw) contains the same error message and someone suggests that they saw this error if the version of `cuda` they've installed is too new for the drivers installed on your system.  I don't know what version of `cuda` you have installed natively on your machine, but maybe `cuda` 12.5 is not compatible with the nVIDIA drivers on your machine?
