### 2024-06-27 PSims/BayesEoR#31 @musoke
That's possible - is cuda 12.5 required for this project? I can try other
versions.
