### 2024-06-27 PSims/BayesEoR#31 @jburba
I don't think so.  We haven't run into any `cuda` version issues in the past.

- Do you have cuda installed on your system already (outside of the bayeseor environment)?
- What NVIDIA driver version do you have currently installed?  That might help us figure out what version of cuda you would need for your system?
