### 2024-06-27 PSims/BayesEoR#31 @musoke
Yes.  `nvidia-smi` says I have cuda version 11.4.

python-cuda version 11.7 works with pytorch in my main project that uses pytorch.
