### 2024-06-27 PSims/BayesEoR#31 @jburba
I might try installing one of those versions of cuda, either 11.4 or 11.7, in your bayeseor conda environment to see if that solves this issue?

It doesn't look like conda-forge hosts cuda 11.x.  You might have to install cuda 11.4 or 11.7 from the nvidia channel by placing
```
  - cuda::nvidia==11.7
```
in your environment.yaml file, for example, if you want cuda 11.7.
