### 2024-07-01 PSims/BayesEoR#31 @musoke
I have tried both that syntax and
```
channels:
  - conda-forge
  - nvidia
```
but was unable to install the older version
