### 2024-07-10 PSims/BayesEoR#31 @jburba
@musoke are you able to try running the test analysis on another machine?  Or is it possible to try updating your drivers for compatibility with later versions of cuda?

From what I can tell, it seems likely these errors are hardware/driver related as opposed to issues with our code base.  It would be nice to know the fix, though, for future users.
