### 2024-07-28 #6667 @dfm
@jpober — Thanks for the ping and sorry for the slow turnaround.

@musoke — Can you take a last look here ASAP? It looks like the BaysEoR developers have responded to your issues. Can we try to get this finished up soon? Thanks!
