### 2024-08-22 PSims/BayesEoR#31 @jburba
Hi @musoke , just checking in to see if you've had a chance to try running the analysis on another machine or updated your drivers to allow for newer versions of cuda?

Alternatively, if you just want to test the analysis, it can run on CPUs only.  It will take longer, however.  For reference, it should take ~4 hours on a single A100 GPU.  We have had someone run the test data on CPUs on a compute cluster.  With 40 cores it ran in ~12 hours.
