### 2024-09-13 PSims/BayesEoR#31 @musoke
Hi @jburba, how do I run a CPU-only analysis? I have run
```
python run-analysis.py --config example-config.yaml --cpu
```
but as far as I can tell from the docs the power spectrum analysis needs GPUs.
