### 2024-09-13 #6667 @musoke
@dfm, thank you for the reminder to come back to this.  From my perspective, all that remains is https://github.com/PSims/BayesEoR/issues/31 - the paper is nice, and I believe that the code does what they claim.
The only thing missing is running the automated tests and such on my own machine.  I have checked with @jburba about how to do the CPU only analysis.
