### 2024-09-29 #6667 @dfm
@musoke — Thanks for this update, and I'm sorry for the slow response! At this point, I think it's ok for you not to run these tests. I think the docs are sufficiently clear about the fact that the software is focused on large-scale GPU compute, and (as you say) @zonca was able to validate the performance. If you're happy with everything else, I think we should proceed. What do you think? Thanks again!!
