### 2024-10-05 c22c93d
Small edits to JOSS manuscript
