+++ joss/paper.md
@@ -52,7 +52,7 @@ where $i$ indexes the spherically-averaged $k$ bins, the sum is performed over a
 
 # Running `BayesEoR`
 
-Running a `BayesEoR` analysis requires an input dataset, a model of the instrument, and set of analysis parameters.  A script is provided with `BayesEoR` for convenience which pre-processes a `pyuvdata`-compatible dataset [@hazelton:2017] of visibilities per baseline, time, and frequency into a one-dimensional data vector, the required form of the input dataset to `BayesEoR`.  As part of the inference, `BayesEoR` forward models the instrument which requires an instrument model containing the primary beam response of a "baseline" (pair of antennas) and the "uv sampling" (the length and orientation of each baseline in the data).  The primary beam response is passed via a configuration file or command line argument.  The uv sampling is generated by the aforementioned convenience script to ensure that the ordering of the baseline in the instrument model matches that in the visibility data vector.  Analysis parameters must also be set by the user to specify file paths to input and output data products and model parameters used to construct the data model (e.g. the number of frequencies and times in the data, the number of voxels in the model Fourier domain cube, the field of view of the sky model).  Note however that these analysis parameters must be chosen carefully based on the data to be analyzed (please see section 2.3 of @burba:2023 for more details on choosing model parameters).  Accordingly, because `BayesEoR` forward models the instrument, we generate a model of the sky as part of our model visibilities calculation.  When the EoR and foregrounds can be adequately described by a sky model with a field of view equal to the width of the primary beam, the memory requirements for a `BayesEoR` analysis are on the order of 10 GB.  This is the case for the provided test dataset for which the disk and RAM requirements are ~17 GB and ~12 GB, respectively.  The field of view for the EoR and foreground sky models can be set independently, however, which allows for the EoR to be modelled within the primary field of view of the telescope while the foregrounds can be modelled across the whole sky.  We wish to note however that, in this fashion, modelling the whole sky can be computationally demanding depending upon the data being analyzed.  For example, in @burba:2023 we show that analyzing a relatively modest dataset (compared to those typically analyzed by HERA, LoFAR, or the MWA) can require ~400 GB of RAM[^2]. Please see section 6.1 of @burba:2023 for more details.
[…]
