### 2024-11-03 #6667 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
