### 2024-11-03 #6667 @dfm
@musoke, @zonca — Many thanks again for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!

@jburba — I have now handed your paper off for a final review by the track editor @warrickball, who may have some final edits before publication. Thanks again for your patience!
