### 2024-11-13 128706e
Editorial tweaks for publication in JOSS
