### 2024-11-13 #6667 @warrickball
Hi @jburba — I've been quietly working through a few checks but I do apologise for the slightly slow response to this. I've opened [a pull request](https://github.com/PSims/BayesEoR/pull/37) with a few final editorial touches. With those implemented I'll test the metadata generation one more time then almost certainly be able to proceed with publication.
