### 2024-11-18 #6667 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
